test: RSpec and Jest suites — 109 backend examples, 88 frontend tests
Backend (109 examples, 100% line coverage):
- spec/models/employee_spec.rb — validations and scopes
- spec/serializers/employee_serializer_spec.rb — field shape and types
- spec/services/employees/{create,update,list}_service_spec.rb — CRUD
logic
- spec/services/insights/salary_insights_service_spec.rb — all 4
aggregation queries
- spec/requests/health_spec.rb — health endpoint
- spec/requests/api/v1/{employees,insights}_spec.rb — full request
coverage

Also fixed PERCENTILE_CONT SQL to cast double precision → numeric for
compatibility with PostgreSQL 12 ROUND(numeric, int) signature.

Frontend (88 tests, 15 test suites):
- lib/format, services/api, hooks/useEmployees + useInsights
- All UI, employee, and insights components
Jest config switched from next/jest SWC transform to babel-jest using
Next.js bundled presets (workaround for @next/swc bus error on this
CPU).

+++ backend/app/services/insights/salary_insights_service.rb
@@ -16,10 +16,10 @@ module Insights
     SALARY_PERCENTILES_SQL = <<~SQL.freeze
       SELECT
         country,
-        ROUND(PERCENTILE_CONT(0.25) WITHIN GROUP (ORDER BY salary), 2) AS p25,
-        ROUND(PERCENTILE_CONT(0.50) WITHIN GROUP (ORDER BY salary), 2) AS p50,
-        ROUND(PERCENTILE_CONT(0.75) WITHIN GROUP (ORDER BY salary), 2) AS p75,
-        ROUND(PERCENTILE_CONT(0.90) WITHIN GROUP (ORDER BY salary), 2) AS p90
+        ROUND(PERCENTILE_CONT(0.25) WITHIN GROUP (ORDER BY salary)::numeric, 2) AS p25,
+        ROUND(PERCENTILE_CONT(0.50) WITHIN GROUP (ORDER BY salary)::numeric, 2) AS p50,
+        ROUND(PERCENTILE_CONT(0.75) WITHIN GROUP (ORDER BY salary)::numeric, 2) AS p75,
+        ROUND(PERCENTILE_CONT(0.90) WITHIN GROUP (ORDER BY salary)::numeric, 2) AS p90
       FROM employees
       GROUP BY country
       ORDER BY country ASC
